+++ package.json
@@ -6,8 +6,7 @@
     "cypress-codegen": "dist/cli.js"
   },
   "files": [
-    "dist",
-    "plugin"
+    "dist"
   ],
   "repository": {
     "type": "git",
@@ -53,7 +52,7 @@
     "vite": "4.4.9"
   },
   "scripts": {
-    "build": "tsc --skipLibCheck && mkdir -p node_modules/cypress-codegen/dist && cp -r dist plugin package.json node_modules/cypress-codegen",
+    "build": "tsc --skipLibCheck && mkdir -p node_modules/cypress-codegen/dist && cp -r dist package.json node_modules/cypress-codegen",
     "cli": "ts-node src/cli.ts",
     "cypress:open": "cypress open",
     "cypress:component": "cypress run --component",
