+++ src/cli.ts
@@ -3,7 +3,7 @@
 /* eslint-disable no-console */
 import * as chalk from 'chalk';
 import { program } from 'commander';
-import { codegen } from './plugin/codegen';
+import { codegen } from './codegen';
 
 console.log(chalk.yellowBright('Generating custom command types...'));
