--- src/plugin/index.ts
@@ -1,41 +1,0 @@
-/*
-Copyright 2022 Expedia, Inc.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    https://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-import { codegen } from './codegen';
-import { join, relative } from 'path';
-
-export type CypressCodegen = (
-  on: Cypress.PluginEvents,
-  config: Cypress.PluginConfigOptions
-) => void;
-
-export const cypressCodegen: CypressCodegen = (
-  on: Cypress.PluginEvents,
-  config: Cypress.PluginConfigOptions
-) => {
-  if (!config.env.skipCodegen) {
-    on('before:browser:launch', async (browser, launchOptions) => {
-      await codegen(config.env.commandsDirectory);
-
-      return launchOptions;
-    });
-  }
-
-  on('task', {
-    getRelativePathToCommands: () => {
-      return relative(join(__dirname, '..'), config.projectRoot);
-    }
-  });
-
-  return config;
-};
